Changed styling on cards
