+++ src/resources/js/Pages/Group/Index.vue
@@ -18,49 +18,20 @@
                     Create New Group
                 </NavLink>
             </div>
-            <div class="grid grid-cols-2 py-12 mt-4 space-y-2 ">
-                <Card v-for="group in props.groups" :key="group.id" class="w-11/12 h-auto py-1 ">
-                    <div class="flex flex-row">
-                        <div class="text-xl font-medium capitalize basis-1/2">
-                            Contact: {{group.contact_firstname}} {{group.contact_lastname}}
-                        </div>
-                        <div class="basis-1/2">
-                            Group Members: {{group.group_size}}
-                        </div>
-                        <div class="grid w-32 h-full grid-cols-2 m-2 space-x-1 text-xl place-items-center ">
-                            <!-- <div class="border m-15 bg-slate-100 hover:cursor-pointer" v-on:click="viewGroup(group.id)">
-                                <EyeIcon/>
-                            </div> -->
-                            <div class="border m-15 bg-slate-100 hover:cursor-pointer" v-on:click="editGroup(group.id)">
-                                <CogIcon/>
-                            </div>
-                            <div class="border m-15 bg-slate-100 hover:cursor-pointer" v-on:click="deleteGroup(group.id)">
-                                <DeleteIcon/>
-                            </div>
-                        </div>
-
-                    </div>
-                </Card>
-
+            <div class="grid grid-cols-4 py-12 mt-4 space-y-2 ">
+                <template v-for="group in props.groups" :key="group.id" class="w-11/12 h-auto py-1 ">
+                    <Card :group="group"/>
+                </template>
             </div>
-
-
         </div>
-
     </AppLayout>
 
 </template>
 
 <script setup>
 import AppLayout from '@/Layouts/AppLayout.vue';
-import Card from '@/Components/Card.vue'
+import Card from '@/Components/Groups/Card.vue'
 import NavLink from '@/Components/NavLink.vue';
-
-import EyeIcon from 'vue-material-design-icons/Eye.vue';
-import CogIcon from 'vue-material-design-icons/Cog.vue';
-import DeleteIcon from 'vue-material-design-icons/Delete.vue';
-
-import { Inertia } from '@inertiajs/inertia';
 
 const props = defineProps({
     groups: Object
@@ -69,13 +40,4 @@ const props = defineProps({
 // const viewGroup = (group) => {
 //         Inertia.get(route('groups.show', group))
 // }
-const editGroup = (group) => {
-        Inertia.get(route('groups.edit', group))
-}
-const deleteGroup = (group) => {
-        Inertia.delete(route('groups.delete', group), {
-            preserveScroll: true
-            }
-        )
-}
 </script>
